Divide MethodInfo class
